+++ lib/cypress_dev/simple_rails_factory.rb
@@ -2,13 +2,14 @@ require 'active_support/core_ext/string'
 
 module CypressDev
   module SimpleRailsFactory
-    def self.create(type, params = {})
-      type.camelize.constantize.create!(params)
+    def self.create(type, *params)
+      params = [{}] if params.empty?
+      type.camelize.constantize.create!(*params)
     end
 
-    def self.create_list(type, amount, params = {})
+    def self.create_list(type, amount, *params)
       amount.to_i.times do
-        create(type,params)
+        create(type,*params)
       end
     end
   end